docs: EXPOSED-313 JSON columns support libraries other than kotlinx.serialization
The current docs and Wiki leverage kotlinx.serialization support before mentioning
the underlying json() that accepts a generic serialization/deserialization combo.
This potentially gives the impression that using other libraries is not supported.

The docs now reflect that the choice of JSON serialization/deserialization library
is left to the user of the module and also includes a new example with Jackson.